pyproject.toml
[project]
name = "inbox-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arcade-ai>=1.1.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
